Terminate Newton method by variation of x
The variation of the y-value (value of determinant) variates depending
on the x-value. For our objective functions, the variation in the
y-value becomes smaller when x is towards zero. It's not suitable to
assign a threshold on the y-value to determine when should terminate
the iteration of the Newton method. Assigned the accepted range of
x-values is better.