Speed up generating expected distance functions
In origin, the recurrence of the expected distance functions is solved
in SR, and the result will be converted to the fraction field of QQ[]
before returning. However, I found directly solving the recurrence of
the expected distance functions in the fraction field of QQ[] is faster
than the original two-step process.