Tool for plotting determinant
Selecting the suitable precision is important. The runtime will be too
long if the precision is too high, and the result will be faulty if the
precision is too low.

One way to know if the precision is high enough to make the results
correct is to plot the determinant. If the curve fluctuates like a
noise, a higher precision should be applied.